+++ src/bin/cmds/doc.rs
@@ -11,24 +11,21 @@ pub fn cli() -> App<'static, 'static> {
     SubCommand::with_name("doc")
         .about("Builds the docs for the root package")
         .arg(args::build_threads())
+        .arg(args::offline())
 }
 
 pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
     let project = current_dir().context(format_err!(
         "couldn't get current dir; doesn't exist or no permissions..."
     ))?;
-    let indices = c.indices.to_vec();
-    let global_cache = c.layout();
-    let logger = logger(c);
-    let threads = match_threads(c, args);
-    let shell = c.shell();
 
     let ctx = build::BuildCtx {
-        indices,
-        global_cache,
-        logger,
-        threads,
-        shell,
+        indices: c.indices.to_vec(),
+        global_cache: c.layout(),
+        logger: logger(c),
+        threads: match_threads(c, args),
+        shell: c.shell(),
+        offline: args.is_present("offline"),
     };
 
     build::doc(&ctx, &project)