+++ src/bin/cmds/test.rs
@@ -12,6 +12,7 @@ pub fn cli() -> App<'static, 'static> {
         .about("Runs the tests of the root package")
         .args(&args::backends())
         .arg(args::build_threads())
+        .arg(args::offline())
         .arg(
             Arg::with_name("test-threads")
                 .long("test-threads")
@@ -29,18 +30,14 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
     let project = current_dir().context(format_err!(
         "couldn't get current dir; doesn't exist or no permissions..."
     ))?;
-    let indices = c.indices.to_vec();
-    let global_cache = c.layout();
-    let logger = logger(c);
-    let threads = match_threads(c, args);
-    let shell = c.shell();
 
     let ctx = build::BuildCtx {
-        indices,
-        global_cache,
-        logger,
-        threads,
-        shell,
+        indices: c.indices.to_vec(),
+        global_cache: c.layout(),
+        logger: logger(c),
+        threads: match_threads(c, args),
+        shell: c.shell(),
+        offline: args.is_present("offline"),
     };
 
     // This is where our default codegen backend is set